:truck: moved types to _typing
